Detect move_self events on Linux
On Linux, if a watched file or directory was moved/renamed watchdog
would not emit any event to the respective handlers. Apparently, this
was by design since the IN_MOVE_SELF event issued by inotify was
being completely ignored. However, when the watch directory or file is
moved on Linux watchdog starts to behave oddly:

  - New events related with that watch are still emitted with the old
  path. Furthermore, if the watch is a directory any created/deleted
  file generates a modified event as well.

  - If the watch is a directory, then every time a file is
  created/deleted a modified event is also emitted.

To avoid this issues, the easiest solution is to handle a self move
event in the same way as a delete self event. Given such an event the
inotify emitter emits a DirectoryDeletedEvent or a FileDeletedEvent,
depending on whether the moved item was a directory or a file.