Detect move_self events on Linux
On Linux, if a watched file or directory was moved/renamed watchdog
would not emit any event to the respective handlers. Apparently, this
was by design since the IN_MOVE_SELF event issued by inotify was
being completely ignored.

With this commit, an IN_MOVE_SELF event cause the inotify emitter to
emit a DirectoryMovedEvent or a FileMovedEvent, depending on whether the
moved item was a directory or a file.